Use curl CURLINFO_CONTENT_LENGTH_DOWNLOAD_T
Previous use of CURLINFO_CONTENT_LENGTH_DOWNLOAD is deprecated so
switch to the updated CURLINFO_CONTENT_LENGTH_DOWNLOAD_T API.

This change does not propagate the curl type used to the rest of
swupd-client but instead tries to use reasonable type casts for their
given purpose.

+++ src/swupd_lib/swupd_curl.h
@@ -6,8 +6,8 @@
  * @brief Wrapper to curl API calls.
  */
 
-#include "swupd_progress.h"
 #include <stdbool.h>
+#include <sys/types.h>
 
 #ifdef __cplusplus
 extern "C" {
@@ -80,7 +80,7 @@ void swupd_curl_cleanup(void);
 /**
  * @brief Query the file content size of 'url' without downloading the full file.
  */
-double swupd_curl_query_content_size(char *url);
+int64_t swupd_curl_query_content_size(char *url);
 
 /**
  * @brief Download @c url and save it in @c filename.
